+++ pybillomat/clients.py
@@ -10,7 +10,7 @@ Clients
 import datetime
 import xml.etree.ElementTree as ET
 import errors
-from bunch import Bunch
+from munch import Munch as Bunch
 from http import Url
 from _items_base import Item, ItemsIterator