+++ pybillomat/credit_note_tags.py
@@ -10,7 +10,7 @@ CreditNote-Tags
 import urllib3
 import xml.etree.ElementTree as ET
 import errors
-from bunch import Bunch
+from munch import Munch as Bunch
 from http import Url
 from _items_base import Item, ItemsIterator