+++ pybillomat/recurrings.py
@@ -11,7 +11,7 @@ Recurrings (Abo-Rechnungen)
 import datetime
 import xml.etree.ElementTree as ET
 import errors
-from bunch import Bunch
+from munch import Munch as Bunch
 from http import Url
 from _items_base import Item, ItemsIterator