+++ pybillomat/suppliers.py
@@ -10,7 +10,7 @@ Suppliers
 
 import errors
 import xml.etree.ElementTree as ET
-from bunch import Bunch
+from munch import Munch as Bunch
 from http import Url
 from _items_base import Item, ItemsIterator